+++ package.json
@@ -18,7 +18,7 @@
   },
   "scripts": {
     "clean": "rm -rf dist",
-    "generate": "yarn clean && ts-node --esm config/build.ts",
+    "generate": "yarn clean && ts-node --esm style-dictionary/build.ts",
     "storybook": "storybook dev -p 6007",
     "test": "echo \"Error: no test specified\" && exit 1",
     "build-storybook": "storybook build"
