+++ style-dictionary/formats/registerJavascriptNested.ts
