+++ style-dictionary/formats/registerJsonFlat.ts
