+++ style-dictionary/formats/registerTypescriptNestedDefinitions.ts
