+++ style-dictionary/utilities/formatProperties.ts
