+++ style-dictionary/utilities/formatValue.ts
