+++ style-dictionary/utilities/generateNestedOutput.ts
