+++ style-dictionary/utilities/generateTypeDeclaration.ts
