+++ style-dictionary/utilities/toPascalCase.ts
